multiplayer code cleanups
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@100

+++ 3dc/Kshape.c
@@ -6343,7 +6343,7 @@ void CreateStarArray(void)
 		StarArray[i].Position.vy = ONE_FIXED-(SeededFastRandom()&131071);
 		{
 			float y = ((float)StarArray[i].Position.vy)/65536.0f;
-			y = sqrt(1-y*y);
+			y = (float)(sqrt(1-y*y));
 
 			f2i(StarArray[i].Position.vx,(float)GetCos(phi)*y);
 			f2i(StarArray[i].Position.vz,(float)GetSin(phi)*y);
